+++ libraries/INA238/CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [0.1.2] - 2025-12-27
+## [0.1.3] - 2026-04-14
+- merge #6, Fix getTemperature() - kudos to CentyLab
+- add alert example
+- add setOverCurrentLimit() + getOverCurrentLimit()
+- minor edits
+
+## [0.1.2] - 2026-03-20
 - fix #4, add **setAddress(newAddress)**
 - minor edits